@@ -1,34 +1,20 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
+export default {
   content: [
-    './src/pages/**/*.{js,ts,jsx,tsx,mdx}',
-    './src/components/**/*.{js,ts,jsx,tsx,mdx}',
-    './src/app/**/*.{js,ts,jsx,tsx,mdx}',
+    "./index.html",
+    "./src/**/*.{js,ts,jsx,tsx}",
   ],
   theme: {
     extend: {
       animation: {
-        scan: 'scan 2.5s cubic-bezier(0.4, 0, 0.6, 1) infinite',
-        glitch: 'glitch 1s linear infinite',
-        float: 'float 6s ease-in-out infinite',
+        'scan': 'scan 2s ease-in-out infinite',
       },
       keyframes: {
         scan: {
-          '0%, 100%': { transform: 'translateY(-200%)' },
-          '50%': { transform: 'translateY(200%)' }
+          '0%, 100%': { transform: 'translateY(-100px)' },
+          '50%': { transform: 'translateY(100px)' },
         },
-        glitch: {
-          '0%, 100%': { transform: 'translate(0)' },
-          '20%': { transform: 'translate(-2px, 2px)' },
-          '40%': { transform: 'translate(-2px, -2px)' },
-          '60%': { transform: 'translate(2px, 2px)' },
-          '80%': { transform: 'translate(2px, -2px)' }
-        },
-        float: {
-          '0%, 100%': { transform: 'translateY(0)' },
-          '50%': { transform: 'translateY(-20px)' }
-        }
-      }
+      },
     },
   },
   plugins: [],
